--- src/card-props.js
@@ -1,6 +1,0 @@
-const colors = [`pink`, `yellow`, `blue`, `black`];
-const modifiers = [`card--deadline`, `card--repeat`, ``];
-const texts = [`Train dog`, `Need to learn JS`, `Sleep well`, `Buy more synths`, `Lorem ipsum`, ``];
-const hashTagsMock = [`#one`, `#one #two`, ``];
-
-export {colors, modifiers, texts, hashTagsMock};
